Form Engine: Integration Branch (2) (#19123)
* VATEAM-90582: Add 21-4140 Digital Form to seed script (#19047)

* Add create_digital_forms

* Add 21-4140 to digital forms script

* Switch includeDob to snake case

* VATEAM-90628: Add Identification Information Paragraph type (#19190)

* Add OMB info fields to Digital Form (#19078)

* Add Identification Information pattern

* Add OMB Info to digital-forms script

* Add Identification Information to digital-forms script

* VATEAM-90650: Create Address Digital Form Pattern (#19331)

Creates a Paragraph type in Drupal for the "Address" pattern. The Paragraph type includes an "Include military address checkbox?" boolean field. Adds the new Paragraph type to the list of available Paragraph types in the "Steps" field of the Digital Form content type.

* VATEAM-90651: Create Phone and Email Digital Form pattern (#19464)

Creates a Paragraph type in Drupal for the "Phone and Email Address" pattern. The Paragraph type includes an "Include email address?" boolean field. Adds the new Paragraph type to the list of available Paragraph types in the "Steps" field of the Digital Form content type.

* VATEAM-94366: Create "Your personal information" Paragraph type (#19534)

Create a new Paragraph type called "Your personal information" that contains the "Name and Date of Birth" and "Identification Information" Paragraph types within it. Remove "Name and Date of Birth" and "Identification Information" as selectable Steps. Update the Digital Forms script to use the new Paragraph type.

* VATEAM-90652: Create List & Loop pattern (#19816)

Creates a Paragraph type in Drupal for the "List & Loop" pattern. The Paragraph type includes an "Optional?" boolean field. Adds the new Paragraph type to the list of available Paragraph types in the "Steps" field of the Digital Form content type.

---------

Co-authored-by: Derek Houck <derek@derekhouck.com>
Co-authored-by: Derek Houck <derek.houck@agile6.com>
Co-authored-by: Derek Houck <12766168+derekhouck@users.noreply.github.com>